Implement auton routine
Major restructuring of DriveTrain, scrapped a bunch of stuff in favor of
using mapped continuous ranges to hopefully add some decent ramping to
the motor control without having to do any full scale motion profiling.

I still probably need to add some manual override so theres no danger
of getting stuck in the auton routine. If possible, after we make sure
this is working I'd like to reimplement it with flags and no while loops
for more stability (and elegance).